feat(profile): redesign division card + fix matchmaking numeric return type
Division card (placed state) redesigned to the new layout: big 2-decimal level
+ league word and week trend, a NEXT/next-milestone ring, a rating-reliability
block with a real "play N more to confirm" count (10 - competitive_matches,
now read through Ranking), progress-to-next-level bar with "+X to go", and an
inset last-ranked-match card with the week's level movement. Drops the old hex
badge / tier segments / week strip / promotion banner. Unranked state
unchanged.

Fix (42804): profiles.level is double precision on the live DB, so
coalesce(rating, level, 2.0) promoted mm_candidates/mm_result_hero outputs to
double and mismatched their numeric columns. Cast the rating/level outputs to
::numeric. Standalone delta matchmaking_fix_numeric_types.sql; folded into the
canonical migration and back-patched the earlier deltas.

+++ supabase/changes/2026-07-15_matchmaking_p1_visibility.sql
@@ -74,7 +74,7 @@ begin
 
   -- Qualify with the table alias: the RETURNS TABLE column `city` shadows an
   -- unqualified `city` here (42702 ambiguous reference otherwise).
-  select coalesce(p.rating, p.level, 2.0), p.city, (coalesce(p.placement_played, 0) < 5)
+  select coalesce(p.rating, p.level, 2.0)::numeric, p.city, (coalesce(p.placement_played, 0) < 5)
     into v_rating, v_city, v_placement
     from public.profiles p where p.id = v_uid;
 
@@ -85,9 +85,9 @@ begin
   select m.id, m.scheduled_at, m.match_type,
          c.name, c.venue_name, coalesce(c.city, cp.city),
          m.created_by, cp.name,
-         coalesce(cp.rating, cp.level, 2.0), coalesce(cp.level, cp.rating, 2.0),
+         coalesce(cp.rating, cp.level, 2.0)::numeric, coalesce(cp.level, cp.rating, 2.0)::numeric,
          (select count(*)::int from public.match_players mp where mp.match_id = m.id),
-         coalesce(m.mm_center_rating, cp.rating, cp.level, 2.0),
+         coalesce(m.mm_center_rating, cp.rating, cp.level, 2.0)::numeric,
          greatest(0, round((1 - abs(v_rating - coalesce(m.mm_center_rating, cp.rating, cp.level, 2.0)) / 3.5) * 100))::int
     from public.matches m
     join public.profiles cp on cp.id = m.created_by
