+++ supabase/migration_player_app.sql
@@ -1596,20 +1596,22 @@ begin
 
   -- Qualify with the table alias: the RETURNS TABLE column `city` shadows an
   -- unqualified `city` here (42702 ambiguous reference otherwise).
-  select coalesce(p.rating, p.level, 2.0), p.city, (coalesce(p.placement_played, 0) < 5)
+  select coalesce(p.rating, p.level, 2.0)::numeric, p.city, (coalesce(p.placement_played, 0) < 5)
     into v_rating, v_city, v_placement
     from public.profiles p where p.id = v_uid;
 
   v_window := coalesce(
     (select value::numeric from public.app_settings where key = 'mm_time_window_hours'), 12);
 
+  -- ::numeric casts: profiles.level is double precision on the live DB, so the
+  -- coalesces would promote to double and mismatch the numeric out-columns (42804).
   return query
   select m.id, m.scheduled_at, m.match_type,
          c.name, c.venue_name, coalesce(c.city, cp.city),
          m.created_by, cp.name,
-         coalesce(cp.rating, cp.level, 2.0), coalesce(cp.level, cp.rating, 2.0),
+         coalesce(cp.rating, cp.level, 2.0)::numeric, coalesce(cp.level, cp.rating, 2.0)::numeric,
          (select count(*)::int from public.match_players mp where mp.match_id = m.id),
-         coalesce(m.mm_center_rating, cp.rating, cp.level, 2.0),
+         coalesce(m.mm_center_rating, cp.rating, cp.level, 2.0)::numeric,
          greatest(0, round((1 - abs(v_rating - coalesce(m.mm_center_rating, cp.rating, cp.level, 2.0)) / 3.5) * 100))::int
     from public.matches m
     join public.profiles cp on cp.id = m.created_by
@@ -1753,10 +1755,10 @@ begin
          mp.team,
          m.score_team_a,
          m.score_team_b,
-         (select rh.delta from public.ranking_history rh
+         (select rh.delta::numeric from public.ranking_history rh
             where rh.profile_id = v_uid and rh.match_id = m.id
             order by rh.created_at desc limit 1),
-         (select rh.rating_after from public.ranking_history rh
+         (select rh.rating_after::numeric from public.ranking_history rh
             where rh.profile_id = v_uid and rh.match_id = m.id
             order by rh.created_at desc limit 1),
          m.match_type
